cassandraClient/requirements.txt
cassandra-driver
psycopg2